requirements.txt
fastapi
uvicorn
# sqlalchemy
sqlalchemy-ext
python-dotenv
requests
pydantic
asyncpg
httpx
numpy